SERVER-1625 Compile error fixed

+++ src/mongo/db/ops/query.cpp
@@ -760,7 +760,7 @@ namespace mongo {
         
         QueryResult *qr = (QueryResult *) result.header();
         qr->cursorId = cursorid;
-        curop.debug().cursorid = ( cursorid == 0 ? -1 : qr->cursorId );
+        curop.debug().cursorid = ( cursorid == 0 ? -1 : (long long)qr->cursorId );
         qr->setResultFlagsToOk();
         // qr->len is updated automatically by appendData()
         curop.debug().responseLength = qr->len;
